test: rework preload page tests
