+++ packages/dashboard-frontend/src/services/factory-location-adapter/index.ts
@@ -15,7 +15,7 @@ import { Location } from 'history';
 
 export interface FactoryLocation {
   readonly searchParams: URLSearchParams;
-  readonly isFullPathUrl: boolean;
+  readonly isHttpLocation: boolean;
   readonly isSshLocation: boolean;
   readonly toString: () => string;
 }
@@ -36,7 +36,7 @@ export class FactoryLocationAdapter implements FactoryLocation {
     this.pathname = sanitizedLocation.pathname;
     this.search = new window.URLSearchParams(sanitizedLocation.search);
 
-    if (FactoryLocationAdapter.isFullPathUrl(sanitizedLocation.pathname)) {
+    if (FactoryLocationAdapter.isHttpLocation(sanitizedLocation.pathname)) {
       this.fullPathUrl = sanitizedLocation.pathname;
       if (sanitizedLocation.search) {
         this.fullPathUrl += sanitizedLocation.search;
@@ -51,7 +51,7 @@ export class FactoryLocationAdapter implements FactoryLocation {
     }
   }
 
-  public static isFullPathUrl(href: string): boolean {
+  public static isHttpLocation(href: string): boolean {
     return /^(https?:\/\/.)[-a-zA-Z0-9@:%._+~#=]{2,}\b([-a-zA-Z0-9@:%_+.~#?&/=]*)$/.test(href);
   }
 
@@ -63,7 +63,7 @@ export class FactoryLocationAdapter implements FactoryLocation {
     return this.search;
   }
 
-  get isFullPathUrl(): boolean {
+  get isHttpLocation(): boolean {
     return this.fullPathUrl !== undefined;
   }
